Removed nobelService, added DataManager, and the data from nobel website
DataManager is our manager for all the data :) It loads the JSON files
and handles the data requests from other parts of the projekt. The
dataManager now has it's own instance of dataFormatter, that is our
class for formatting the data to match the D3 charts. So we can now
request data from dataManager and let it format the data to our liking.
Removed the nobelService because we don't need it anymore as we load
json files instead of using an API